requirements.txt
fastapi
uvicorn[standard]
numpy
scipy
faster-whisper
httpx
python-dotenv
claude-code-sdk
